:truck: Move readConfig into config
