+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export * from './defaultConfig';
+export * from './getConfigfile';
+export * from './getValueFromConfig';
